requirements.txt
asgiref==3.4.1
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
dj-database-url==2.0.0
Django==3.2.8
django-cors-headers==3.11.0
django-environ==0.10.0
django-phonenumber-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.1
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
packaging==23.0
phonenumberslite==8.12.45
Pillow==8.3.2
pluggy==1.0.0
psycopg2-binary==2.9.6
PyJWT==2.7.0
pytest==7.2.2
python-dotenv==1.0.0
pytz==2021.3
requests==2.28.2
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==3.10.0.2
urllib3==1.26.15
whitenoise==6.5.0
wonderwords==2.2.0
